fix: supabase-admin and API route refactor.
