+++ pages/api/create-portal-link.ts
@@ -1,15 +1,16 @@
 import { stripe } from 'utils/stripe';
-import { getUser } from 'utils/supabase-admin';
-import { createOrRetrieveCustomer } from 'utils/useDatabase';
+import {
+  getUser,
+  withAuthRequired
+} from '@supabase/supabase-auth-helpers/nextjs';
+import { createOrRetrieveCustomer } from 'utils/supabase-admin';
 import { getURL } from 'utils/helpers';
 import { NextApiRequest, NextApiResponse } from 'next';
 
 const createPortalLink = async (req: NextApiRequest, res: NextApiResponse) => {
   if (req.method === 'POST') {
-    const token = req.headers.token as string;
-
     try {
-      const user = await getUser(token);
+      const { user } = await getUser({ req, res });
       if (!user) throw Error('Could not get user');
       const customer = await createOrRetrieveCustomer({
         uuid: user.id || '',
@@ -25,7 +26,9 @@ const createPortalLink = async (req: NextApiRequest, res: NextApiResponse) => {
       return res.status(200).json({ url });
     } catch (err: any) {
       console.log(err);
-      res.status(500).json({ error: { statusCode: 500, message: err.message } });
+      res
+        .status(500)
+        .json({ error: { statusCode: 500, message: err.message } });
     }
   } else {
     res.setHeader('Allow', 'POST');
@@ -33,4 +36,4 @@ const createPortalLink = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 };
 
-export default createPortalLink;
+export default withAuthRequired(createPortalLink);
