+++ utils/supabase-admin.ts
@@ -1,16 +1,179 @@
 import { createClient } from '@supabase/supabase-js';
+import { stripe } from './stripe';
+import { toDateTime } from './helpers';
+import { Customer, UserDetails, Price, Product } from 'types';
+import Stripe from 'stripe';
 
-export const supabaseAdmin = createClient(
+// Note: supabaseAdmin uses the SERVICE_ROLE_KEY which you must only use in a secure server-side context
+// as it has admin priviliges and overwrites RLS policies!
+const supabaseAdmin = createClient(
   process.env.NEXT_PUBLIC_SUPABASE_URL || '',
   process.env.SUPABASE_SERVICE_ROLE_KEY || ''
 );
 
-export const getUser = async (token: string) => {
-  const { data, error } = await supabaseAdmin.auth.api.getUser(token);
+const upsertProductRecord = async (product: Stripe.Product) => {
+  const productData: Product = {
+    id: product.id,
+    active: product.active,
+    name: product.name,
+    description: product.description ?? undefined,
+    image: product.images?.[0] ?? null,
+    metadata: product.metadata
+  };
 
+  const { error } = await supabaseAdmin
+    .from<Product>('products')
+    .upsert([productData]);
+  if (error) throw error;
+  console.log(`Product inserted/updated: ${product.id}`);
+};
+
+const upsertPriceRecord = async (price: Stripe.Price) => {
+  const priceData: Price = {
+    id: price.id,
+    product_id: typeof price.product === 'string' ? price.product : '',
+    active: price.active,
+    currency: price.currency,
+    description: price.nickname ?? undefined,
+    type: price.type,
+    unit_amount: price.unit_amount ?? undefined,
+    interval: price.recurring?.interval,
+    interval_count: price.recurring?.interval_count,
+    trial_period_days: price.recurring?.trial_period_days,
+    metadata: price.metadata
+  };
+
+  const { error } = await supabaseAdmin
+    .from<Price>('prices')
+    .upsert([priceData]);
+  if (error) throw error;
+  console.log(`Price inserted/updated: ${price.id}`);
+};
+
+const createOrRetrieveCustomer = async ({
+  email,
+  uuid
+}: {
+  email: string;
+  uuid: string;
+}) => {
+  const { data, error } = await supabaseAdmin
+    .from<Customer>('customers')
+    .select('stripe_customer_id')
+    .eq('id', uuid)
+    .single();
   if (error) {
-    throw error;
+    // No customer record found, let's create one.
+    const customerData: { metadata: { supabaseUUID: string }; email?: string } =
+      {
+        metadata: {
+          supabaseUUID: uuid
+        }
+      };
+    if (email) customerData.email = email;
+    const customer = await stripe.customers.create(customerData);
+    // Now insert the customer ID into our Supabase mapping table.
+    const { error: supabaseError } = await supabaseAdmin
+      .from('customers')
+      .insert([{ id: uuid, stripe_customer_id: customer.id }]);
+    if (supabaseError) throw supabaseError;
+    console.log(`New customer created and inserted for ${uuid}.`);
+    return customer.id;
   }
+  if (data) return data.stripe_customer_id;
+};
 
-  return data;
+/**
+ * Copies the billing details from the payment method to the customer object.
+ */
+const copyBillingDetailsToCustomer = async (
+  uuid: string,
+  payment_method: Stripe.PaymentMethod
+) => {
+  //Todo: check this assertion
+  const customer = payment_method.customer as string;
+  const { name, phone, address } = payment_method.billing_details;
+  if (!name || !phone || !address) return;
+  //@ts-ignore
+  await stripe.customers.update(customer, { name, phone, address });
+  const { error } = await supabaseAdmin
+    .from<UserDetails>('users')
+    .update({
+      billing_address: address,
+      payment_method: payment_method[payment_method.type]
+    })
+    .eq('id', uuid);
+  if (error) throw error;
+};
+
+const manageSubscriptionStatusChange = async (
+  subscriptionId: string,
+  customerId: string,
+  createAction = false
+) => {
+  // Get customer's UUID from mapping table.
+  const { data: customerData, error: noCustomerError } = await supabaseAdmin
+    .from<Customer>('customers')
+    .select('id')
+    .eq('stripe_customer_id', customerId)
+    .single();
+  if (noCustomerError) throw noCustomerError;
+
+  const { id: uuid } = customerData || {};
+
+  const subscription = await stripe.subscriptions.retrieve(subscriptionId, {
+    expand: ['default_payment_method']
+  });
+  // Upsert the latest status of the subscription object.
+  const subscriptionData = {
+    id: subscription.id,
+    user_id: uuid,
+    metadata: subscription.metadata,
+    status: subscription.status,
+    price_id: subscription.items.data[0].price.id,
+    //TODO check quantity on subscription
+    // @ts-ignore
+    quantity: subscription.quantity,
+    cancel_at_period_end: subscription.cancel_at_period_end,
+    cancel_at: subscription.cancel_at
+      ? toDateTime(subscription.cancel_at)
+      : null,
+    canceled_at: subscription.canceled_at
+      ? toDateTime(subscription.canceled_at)
+      : null,
+    current_period_start: toDateTime(subscription.current_period_start),
+    current_period_end: toDateTime(subscription.current_period_end),
+    created: toDateTime(subscription.created),
+    ended_at: subscription.ended_at ? toDateTime(subscription.ended_at) : null,
+    trial_start: subscription.trial_start
+      ? toDateTime(subscription.trial_start)
+      : null,
+    trial_end: subscription.trial_end
+      ? toDateTime(subscription.trial_end)
+      : null
+  };
+
+  const { error } = await supabaseAdmin
+    .from('subscriptions')
+    .upsert([subscriptionData]);
+  if (error) throw error;
+  console.log(
+    `Inserted/updated subscription [${subscription.id}] for user [${uuid}]`
+  );
+
+  // For a new subscription copy the billing details to the customer object.
+  // NOTE: This is a costly operation and should happen at the very end.
+  if (createAction && subscription.default_payment_method && uuid)
+    //@ts-ignore
+    await copyBillingDetailsToCustomer(
+      uuid,
+      subscription.default_payment_method as Stripe.PaymentMethod
+    );
+};
+
+export {
+  upsertProductRecord,
+  upsertPriceRecord,
+  createOrRetrieveCustomer,
+  manageSubscriptionStatusChange
 };
